pyproject.toml
[project]
name = "bic-test-repository-llm"
version = "0.1.0"
description = "Test work"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.0",
    "itsdangerous>=2.2.0",
    "pydantic-settings>=2.10.1",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
venvPath = "."
venv = ".venv"

[dependency-groups]
lint = [
    "pylint>=3.3.7",
    "ruff>=0.12.2",
]
test = [
    "pytest>=8.4.1"
]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
